common: moving ArithTraits to KokkosKernels_ArithTraits.hpp
This avoid having us use the naming convention of Kokkos Core
for something that lives in Kokkos Kernels.
Next, I will modify header inclusions in the library to reflect
this change.
